Continuing with modularizing torch_training_utils
