+++ src/yoke/torch_training_utils.py
@@ -14,31 +14,6 @@ import torch
 import torch.distributed as dist
 
 
-####################################
-# Saving Results
-####################################
-def save_append_df(path: str, df: pd.DataFrame, START: bool):
-    """Function to save/append dataframe contents to a csv file
-
-    Args:
-        path (str): path of csv file
-        df (pd.DataFrame): pandas dataframe to save
-        START (bool): indicates if the file path needs to be initiated
-
-    Returns:
-        No Return Objects
-
-    """
-    if START:
-        assert not os.path.isfile(path), (
-            "If starting training, " + path + " should not exist."
-        )
-        df.to_csv(path, header=True, index=True, mode="x")
-    else:
-        assert os.path.isfile(path), "If continuing training, " + path + " should exist."
-        df.to_csv(path, header=False, index=True, mode="a")
-
-
 def append_to_dict(dictt: dict, batch_ID: int, truth, pred, loss):
     """Function to appending sample information to a dictionary Dictionary must
     be initialized with correct keys
@@ -69,90 +44,6 @@ def append_to_dict(dictt: dict, batch_ID: int, truth, pred, loss):
 ####################################
 # Training on a Datastep
 ####################################
-def train_scalar_datastep(
-        data: tuple,
-        model: torch.nn.Module,
-        optimizer: torch.optim.Optimizer,
-        loss_fn: torch.nn.Module,
-        device: torch.device) -> tuple:
-    """Function to complete a training step on a single sample.
-
-    For this training step the network's output is a scalar.
-
-    Args:
-        data (tuple): tuple of model input and corresponding ground truth
-        model (loaded pytorch model): model to train
-        optimizer (torch.optim): optimizer for training set
-        loss_fn (torch.nn Loss Function): loss function for training set
-        device (torch.device): device index to select
-
-    Returns:
-        loss (): evaluated loss for the data sample
-
-    """
-    # Set model to train
-    model.train()
-
-    # Extract data
-    (inpt, truth) = data
-    inpt = inpt.to(device, non_blocking=True)
-    # Unsqueeze is necessary for scalar ground-truth output
-    truth = truth.to(torch.float32).unsqueeze(-1).to(device, non_blocking=True)
-
-    # Perform a forward pass
-    # NOTE: If training on GPU model should have already been moved to GPU
-    # prior to initalizing optimizer.
-    pred = model(inpt)
-    loss = loss_fn(pred, truth)
-
-    # Perform backpropagation and update the weights
-    optimizer.zero_grad(set_to_none=True)
-    loss.mean().backward()
-    optimizer.step()
-
-    return truth, pred, loss
-
-
-def train_array_datastep(data: tuple, model, optimizer, loss_fn, device: torch.device):
-    """Function to complete a training step on a single sample in which the
-    network's output is an array.
-
-    Args:
-        data (tuple): tuple of model input and corresponding ground truth
-        model (loaded pytorch model): model to train
-        optimizer (torch.optim): optimizer for training set
-        loss_fn (torch.nn Loss Function): loss function for training set
-        device (torch.device): device index to select
-
-    Returns:
-        loss (): evaluated loss for the data sample
-
-    """
-    # Set model to train
-    model.train()
-
-    # Extract data
-    (inpt, truth) = data
-    inpt = inpt.to(device, non_blocking=True)
-    truth = truth.to(device, non_blocking=True)
-
-    # Perform a forward pass
-    # NOTE: If training on GPU model should have already been moved to GPU
-    # prior to initalizing optimizer.
-    pred = model(inpt)
-    loss = loss_fn(pred, truth)
-
-    per_sample_loss = loss.mean(dim=[1, 2, 3])  # Shape: (batch_size,)
-
-    # Perform backpropagation and update the weights
-    # optimizer.zero_grad()
-    optimizer.zero_grad(set_to_none=True)  # Possible speed-up
-    loss.mean().backward()
-    optimizer.step()
-
-    return truth, pred, per_sample_loss
-
-
 def train_loderunner_datastep(
         data: tuple,
         model,
@@ -372,146 +263,9 @@ def train_DDP_loderunner_datastep(
     return end_img, pred_img, all_losses
 
 
-def train_loderunner_fabric_datastep(
-        fabric,
-        data: tuple,
-        model,
-        optimizer,
-        loss_fn):
-    """A training step for which the data is of multi-input, multi-output type.
-
-    This is currently a proto-type function to get the LodeRunner architecture
-    training on a non-variable set of channels.
-
-    Args:
-        fabric (lightning.fabric.Fabric): Fabric instance.
-        data (tuple): tuple of model input, corresponding ground truth, and lead time
-        model (loaded pytorch model): model to train
-        optimizer (torch.optim): optimizer for training set
-        loss_fn (torch.nn Loss Function): loss function for training set
-
-    Returns:
-        loss (): evaluated loss for the data sample
-
-    """
-    # Set model to train
-    model.train()
-
-    # Distribute with fabric
-    start_img, end_img, Dt = data
-
-    # For our first LodeRunner training on the lsc240420 dataset the input and
-    # output prediction variables are fixed.
-    #
-    # Both in_vars and out_vars correspond to indices for every variable in
-    # this training setup...
-    #
-    # in_vars = ['density_case',
-    #            'density_cushion',
-    #            'density_maincharge',
-    #            'density_outside_air',
-    #            'density_striker',
-    #            'density_throw',
-    #            'Uvelocity',
-    #            'Wvelocity']
-    in_vars = torch.tensor([0, 1, 2, 3, 4, 5, 6, 7])
-    in_vars = fabric.to_device(in_vars)
-    out_vars = torch.tensor([0, 1, 2, 3, 4, 5, 6, 7])
-    out_vars = fabric.to_device(out_vars)
-
-    # Perform a forward pass
-    # NOTE: If training on GPU model should have already been moved to GPU
-    # prior to initalizing optimizer.
-    pred_img = model(start_img, in_vars, out_vars, Dt)
-
-    # Expecting to use a *reduction="none"* loss function so we can track loss
-    # between individual samples. However, this will make the loss be computed
-    # element-wise so we need to still average over the (channel, height,
-    # width) dimensions to get the per-sample loss.
-    loss = loss_fn(pred_img, end_img)
-    per_sample_loss = loss.mean(dim=[1, 2, 3])  # Shape: (batch_size,)
-
-    # Perform backpropagation and update the weights
-    optimizer.zero_grad(set_to_none=True)  # Possible speed-up
-    fabric.backward(loss.mean())
-    optimizer.step()
-
-    # Gather per-sample loss across all processes
-    global_per_sample_loss = fabric.all_gather(per_sample_loss)
-    global_per_sample_loss = global_per_sample_loss.flatten()
-
-    # Delete created tensors to free memory
-    del in_vars
-    del out_vars
-
-    # Clear GPU memory after each deallocation
-    torch.cuda.empty_cache()
-
-    return end_img, pred_img, global_per_sample_loss
-
-
 ####################################
 # Evaluating on a Datastep
 ####################################
-def eval_scalar_datastep(data: tuple, model, loss_fn, device: torch.device):
-    """Function to complete a validation step on a single sample for which the
-    network output is a scalar.
-
-    Args:
-        data (tuple): tuple of model input and corresponding ground truth
-        model (loaded pytorch model): model evaluate
-        loss_fn (torch.nn Loss Function): loss function for training set
-        device (torch.device): device index to select
-
-    Returns:
-        loss (): evaluated loss for the data sample
-
-    """
-    # Set model to eval
-    model.eval()
-
-    # Extract data
-    (inpt, truth) = data
-    inpt = inpt.to(device, non_blocking=True)
-    truth = truth.to(torch.float32).unsqueeze(-1).to(device, non_blocking=True)
-
-    # Perform a forward pass
-    pred = model(inpt)
-    loss = loss_fn(pred, truth)
-
-    return truth, pred, loss
-
-
-def eval_array_datastep(data: tuple, model, loss_fn, device: torch.device):
-    """Function to complete a validation step on a single sample in which network
-    output is an array.
-
-    Args:
-        data (tuple): tuple of model input and corresponding ground truth
-        model (loaded pytorch model): model evaluate
-        loss_fn (torch.nn Loss Function): loss function for training set
-        device (torch.device): device index to select
-
-    Returns:
-        loss (): evaluated loss for the data sample
-
-    """
-    # Set model to eval
-    model.eval()
-
-    # Extract data
-    (inpt, truth) = data
-    inpt = inpt.to(device, non_blocking=True)
-    truth = truth.to(device, non_blocking=True)
-
-    # Perform a forward pass
-    pred = model(inpt)
-    loss = loss_fn(pred, truth)
-    per_sample_loss = loss.mean(dim=[1, 2, 3])  # Shape: (batch_size,)
-
-    return truth, pred, per_sample_loss
-
-
 def eval_loderunner_datastep(
         data: tuple,
         model,
@@ -714,78 +468,6 @@ def eval_DDP_loderunner_datastep(
     torch.cuda.empty_cache()
 
     return end_img, pred_img, all_losses
-
-
-def eval_loderunner_fabric_datastep(
-        fabric,
-        data: tuple,
-        model,
-        loss_fn):
-    """An evaluation step for which the data is of multi-input, multi-output type.
-
-    This is currently a proto-type function to get the LodeRunner architecture
-    training on a non-variable set of channels.
-
-    Args:
-        fabric (lightning.fabric.Fabric): Fabric instance.
-        data (tuple): tuple of model input, corresponding ground truth, and lead time
-        model (loaded pytorch model): model to train
-        loss_fn (torch.nn Loss Function): loss function for training set
-        device (torch.device): device index to select
-
-    Returns:
-        loss (): evaluated loss for the data sample
-
-    """
-    # Set model to train
-    model.eval()
-
-    # Extract data
-    start_img, end_img, Dt = data
-
-    # For our first LodeRunner training on the lsc240420 dataset the input and
-    # output prediction variables are fixed.
-    #
-    # Both in_vars and out_vars correspond to indices for every variable in
-    # this training setup...
-    #
-    # in_vars = ['density_case',
-    #            'density_cushion',
-    #            'density_maincharge',
-    #            'density_outside_air',
-    #            'density_striker',
-    #            'density_throw',
-    #            'Uvelocity',
-    #            'Wvelocity']
-    in_vars = torch.tensor([0, 1, 2, 3, 4, 5, 6, 7])
-    in_vars = fabric.to_device(in_vars)
-    out_vars = torch.tensor([0, 1, 2, 3, 4, 5, 6, 7])
-    out_vars = fabric.to_device(out_vars)
-
-    # Perform a forward pass
-    # NOTE: If training on GPU model should have already been moved to GPU
-    # prior to initalizing optimizer.
-    pred_img = model(start_img, in_vars, out_vars, Dt)
-
-    # Expecting to use a *reduction="none"* loss function so we can track loss
-    # between individual samples. However, this will make the loss be computed
-    # element-wise so we need to still average over the (channel, height,
-    # width) dimensions to get the per-sample loss.
-    loss = loss_fn(pred_img, end_img)
-    per_sample_loss = loss.mean(dim=[1, 2, 3])  # Shape: (batch_size,)
-
-    # Gather per-sample loss across all processes
-    global_per_sample_loss = fabric.all_gather(per_sample_loss)
-    global_per_sample_loss = global_per_sample_loss.flatten()
-
-    # Delete created tensors to free memory
-    del in_vars
-    del out_vars
-
-    # Clear GPU memory after each deallocation
-    torch.cuda.empty_cache()
-
-    return end_img, pred_img, global_per_sample_loss
 
 
 ######################################
@@ -1478,103 +1160,6 @@ def train_DDP_loderunner_epoch(
                     torch.cuda.empty_cache()
 
 
-def train_fabric_loderunner_epoch(
-    fabric,
-    training_data,
-    validation_data,
-    model,
-    optimizer,
-    loss_fn,
-    LRsched,
-    epochIDX,
-    train_per_val,
-    train_rcrd_filename: str,
-    val_rcrd_filename: str,
-):
-    """Function to complete a training epoch on the LodeRunner architecture with
-    fixed channels in the input and output. Training and validation information
-    is saved to successive CSV files.
-
-    Args:
-        fabric (lightning.fabric.Fabric): Fabric instance to take care of distributed
-                                          data-parallel training
-        training_data (torch.dataloader): dataloader containing the training samples
-        validation_data (torch.dataloader): dataloader containing the validation samples
-        model (loaded pytorch model): model to train
-        optimizer (torch.optim): optimizer for training set
-        loss_fn (torch.nn Loss Function): loss function for training set
-        LRsched (torch.optim.lr_scheduler): Learning-rate scheduler that will be called
-                                            every training step.
-        epochIDX (int): Index of current training epoch
-        train_per_val (int): Number of Training epochs between each validation
-        train_rcrd_filename (str): Name of CSV file to save training sample stats to
-        val_rcrd_filename (str): Name of CSV file to save validation sample stats to
-
-    """
-    # Initialize things to save
-    trainbatch_ID = 0
-    valbatch_ID = 0
-
-    train_rcrd_filename = train_rcrd_filename.replace("<epochIDX>", f"{epochIDX:04d}")
-    # Train on all training samples
-    with open(train_rcrd_filename, "a") as train_rcrd_file:
-        for traindata in training_data:
-            trainbatch_ID += 1
-
-            truth, pred, train_losses = train_loderunner_fabric_datastep(
-                fabric, traindata, model, optimizer, loss_fn,
-            )
-
-            # Increment the learning-rate scheduler
-            LRsched.step()
-
-            if fabric.global_rank == 0:
-                # Stack loss record and write using numpy
-                batch_records = np.column_stack([
-                    np.full(len(train_losses), epochIDX),
-                    np.full(len(train_losses), trainbatch_ID),
-                    train_losses.detach().cpu().numpy().flatten()
-                ])
-                np.savetxt(train_rcrd_file, batch_records, fmt="%d, %d, %.8f")
-
-            # Explictly delete produced tensors to free memory
-            del truth
-            del pred
-            del train_losses
-
-            # Clear GPU memory after each batch
-            torch.cuda.empty_cache()
-
-    # Evaluate on all validation samples
-    if epochIDX % train_per_val == 0:
-        print("Validating...", epochIDX)
-        val_rcrd_filename = val_rcrd_filename.replace("<epochIDX>", f"{epochIDX:04d}")
-        with open(val_rcrd_filename, "a") as val_rcrd_file:
-            with torch.no_grad():
-                for valdata in validation_data:
-                    valbatch_ID += 1
-                    truth, pred, val_losses = eval_loderunner_fabric_datastep(
-                        fabric, valdata, model, loss_fn,
-                    )
-
-                    if fabric.global_rank == 0:
-                        # Stack loss record and write using numpy
-                        batch_records = np.column_stack([
-                            np.full(len(val_losses), epochIDX),
-                            np.full(len(val_losses), valbatch_ID),
-                            val_losses.detach().cpu().numpy().flatten()
-                        ])
-                        np.savetxt(val_rcrd_file, batch_records, fmt="%d, %d, %.8f")
-
-                    # Explictly delete produced tensors to free memory
-                    del truth
-                    del pred
-                    del val_losses
-
-                    # Clear GPU memory after each batch
-                    torch.cuda.empty_cache()
-
-
 def train_lsc_policy_datastep(
     data: tuple,
     model,
@@ -1609,7 +1194,7 @@ def train_lsc_policy_datastep(
     # Forward pass
     pred_distribution = model(state_y, stateH, targetH)
     pred_mean = pred_distribution.mean
-    
+
     # Compute loss
     loss = loss_fn(pred_mean, x_true)
     per_sample_loss = loss.mean(dim=1)  # Per-sample loss
@@ -1664,7 +1249,7 @@ def eval_lsc_policy_datastep(
     # Forward pass
     with torch.no_grad():
         pred_distribution = model(state_y, stateH, targetH)
-        
+
     pred_mean = pred_distribution.mean
 
     # Compute loss
