+++ src/yoke/utils/training/__init__.py
@@ -1,0 +1,1 @@
+"""Modules for specific training algorithms and utilities."""
